Moves provider tests
