+++ packages/provider/__tests__/article-provider.test.js
@@ -3,8 +3,8 @@ import renderer from "react-test-renderer";
 import { MockedProvider } from "@times-components/utils";
 import { addTypenameToDocument } from "apollo-utilities";
 import fixture from "@times-components/provider-test-tools/fixtures/article.json";
-import { ArticleProvider } from "../provider";
-import { query as articleQuery } from "../article";
+import { ArticleProvider } from "../src/provider";
+import { query as articleQuery } from "../src/article";
 
 const mocks = [
   {
