+++ packages/provider/__tests__/author-articles-with-images.test.js
@@ -2,27 +2,29 @@ import React from "react";
 import renderer from "react-test-renderer";
 import { MockedProvider } from "@times-components/utils";
 import { fixtureGenerator } from "@times-components/provider-test-tools";
-import { AuthorArticlesNoImagesProvider } from "../provider";
+import { AuthorArticlesWithImagesProvider } from "../src/provider";
 
 const renderComponent = child =>
   renderer.create(
     <MockedProvider
-      mocks={fixtureGenerator.makeArticleMocks({ pageSize: 5, delay: 0 })}
+      mocks={fixtureGenerator.makeArticleMocks({
+        withImages: true,
+        pageSize: 5,
+        delay: 0
+      })}
     >
-      <AuthorArticlesNoImagesProvider
+      <AuthorArticlesWithImagesProvider
         slug="deborah-haynes"
         pageSize={5}
         page={1}
-        shortSummaryLength={220}
-        longSummaryLength={360}
         debounceTimeMs={0}
       >
         {child}
-      </AuthorArticlesNoImagesProvider>
+      </AuthorArticlesWithImagesProvider>
     </MockedProvider>
   );
 
-describe("AuthorArticlesNoImagesProvider", () => {
+describe("AuthorArticlesWithImagesProvider", () => {
   it("returns query result", done => {
     renderComponent(({ isLoading, author }) => {
       if (!isLoading) {
